requirements.txt
DateTime==5.1
Flask==2.2.3
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
protobuf==3.20.3
PyJWT==2.6.0
PyMySQL==1.0.3
pytz==2023.3
SQLAlchemy==2.0.9
typing_extensions==4.5.0
uuid==1.30
Werkzeug==2.2.3
zope.interface==6.0